[FW] Corrected EEPROM value cast problems

+++ FW/master/eeprom_map.h
@@ -46,8 +46,8 @@ private:
 public:
     EEPROM_MAP_OBJ();   // Constructor
     ~EEPROM_MAP_OBJ();  // Destructor
-    uint32_t eeprom_get_time();
-    void eeprom_set_time(uint32_t time);
+    uint8_t eeprom_get_time();
+    void eeprom_set_time(uint8_t time);
     void eeprom_write(int add, uint8_t val);
     void reset_eeprom_time();
 };
